code/requirements.txt
speech_recognition
RPi.GPIO